# perf(gatsby-plugin-mdx): drop another babel step during sourcing