This Babel step was used to get exports and remark header but we can use the mdast for this purpose, just like for imports.

> This cuts off **20 to 30%** of the total runtime of a baseline mdx benchmark!

---

## The numbers

The following numbers are comparing this branch vs its base branch (~gatsby@2.24.2) on a dedicated nuc. The benchmark site will randomly generate the requested number of page files, will all look something like this:

```mdx
---
articleNumber: 119
title: "Accusamus non dignissimos sed maxime autem."
path: 'accusamus-non-dignissimos-sed-maxime-autem.'
date: 2018-03-05
---

import { Link } from "gatsby"

export const author = "Fred Flintstone"
export default props => <main {...props} />

<Link to="/">Go Home</Link>

Officia modi et qui ... <truncated paragraph of text>
 
Et eos eum sint laudantium ...  <truncated paragraph of text>
```

Based on that here are the numbers (in truncated seconds):

```
"Done building" time:

     |  PR      | Base     | Delta | % Saved
------------------------------------|--------
  1k |   40     |    51     |   -11 | 21%
  2k |   65     |    85     |   -20 | 23%
  4k |  113     |   153     |   -40 | 26%
  8k |  218     |   316     |   -98 | 31%
 16k |  451     |   651     |  -200 | 30%
 32k | 1094     |  1551     |  -457 | 29%
 64k | 2850     |  3927     | -1077 | 27%

Sourcing time:

     |  PR    | % of total  | Base   | Delta   | % saved (of source time)
--------------------------------------------- ------------------------------
  1k |    4   | 10%         |   16   |   -12   | 75%
  2k |    8   | 12%         |   29   |   -21   | 72%
  4k |   14   | 12%         |   55   |   -41   | 74%
  8k |   30   | 13%         |  116   |   -86   | 74%
 16k |   70   | 15%         |  249   |  -179   | 71%
 32k |  207   | 18%         |  613   |  -408   | 66%
 64k |  707   | 24%         | 1545   |  -838   | 54%
```