@pvdz 2020-07-30
This PR was rebverted due to breaking undocumented API's that were depended on by third parties anyways. That was fixed and the gist of this PR was merged again in #26002